Fixing an overflow warning.

+++ src/theory/arith/theory_arith_private.cpp
@@ -1632,9 +1632,9 @@ bool TheoryArithPrivate::solveRealRelaxation(Theory::Effort effortLevel){
 
     int16_t oldCap = options::arithStandardCheckVarOrderPivots();
 
-    static const int16_t pass2Limit = 10;
+    static const int32_t pass2Limit = 10;
     static const int32_t relaxationLimit = 1000000;
-    static const int16_t mipLimit = 100000;
+    static const int32_t mipLimit = 100000;
 
     d_qflraStatus = simplex.findModel(false);
